requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-dotenv>=1.0.0
supabase>=2.7.0
openai>=1.50.0
anthropic>=0.40.0
python-multipart>=0.0.12
itsdangerous>=2.2.0
httpx>=0.27.0
pydantic-settings>=2.0.0
